- removed additional validation - removed Client class - removed client test and replaced with proper integration test - added retries
