+++ kafka-ui-api/src/test/java/com/provectus/kafka/ui/KafkaConnectServiceTests.java
@@ -1,0 +1,135 @@
+package com.provectus.kafka.ui;
+
+import com.provectus.kafka.ui.model.*;
+import lombok.extern.log4j.Log4j2;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.reactive.AutoConfigureWebTestClient;
+import org.springframework.core.ParameterizedTypeReference;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+import java.util.Map;
+import java.util.UUID;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+@ContextConfiguration(initializers = {AbstractBaseTest.Initializer.class})
+@Log4j2
+@AutoConfigureWebTestClient(timeout = "60000")
+public class KafkaConnectServiceTests extends AbstractBaseTest {
+    private final String clusterName = "local";
+    private final String connectorName = UUID.randomUUID().toString();
+
+    @Autowired
+    private WebTestClient webTestClient;
+
+
+    @BeforeEach
+    public void setUp() {
+        webTestClient.post()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors", clusterName)
+                .bodyValue(new NewConnector()
+                        .name(connectorName)
+                        .config(Map.of(
+                                "connector.class", "org.apache.kafka.connect.file.FileStreamSinkConnector",
+                                "tasks.max", "1",
+                                "topics", "output-topic",
+                                "file", "/tmp/test"
+                        ))
+                )
+                .exchange()
+                .expectStatus().isOk();
+    }
+
+    @AfterEach
+    public void tearDown() {
+        webTestClient.delete()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors/{connectorName}", clusterName, connectorName)
+                .exchange()
+                .expectStatus().isOk();
+    }
+
+    @Test
+    public void shouldListConnectors() {
+        webTestClient.get()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors", clusterName)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                .jsonPath(String.format("$[?(@ == '%s')]", connectorName))
+                .exists();
+    }
+
+    @Test
+    public void shouldRetrieveConnectorStatus() {
+        webTestClient.get()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors/{connectorName}/status", clusterName, connectorName)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody()
+                .jsonPath("$.connector.state")
+                .value(connectorState -> assertEquals(ConnectorStatusConnector.StateEnum.RUNNING.getValue(), connectorState));
+    }
+
+    @Test
+    public void shouldUpdateConfig() {
+        webTestClient.put()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors/{connectorName}/config", clusterName, connectorName)
+                .bodyValue(Map.of(
+                        "connector.class", "org.apache.kafka.connect.file.FileStreamSinkConnector",
+                        "tasks.max", "1",
+                        "topics", "another-topic",
+                        "file", "/tmp/new"
+                        )
+                )
+                .exchange()
+                .expectStatus().isOk();
+
+        webTestClient.get()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connectors/{connectorName}/config", clusterName, connectorName)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(new ParameterizedTypeReference<Map<String, Object>>() {
+                })
+                .isEqualTo(Map.of(
+                        "connector.class", "org.apache.kafka.connect.file.FileStreamSinkConnector",
+                        "tasks.max", "1",
+                        "topics", "another-topic",
+                        "file", "/tmp/new",
+                        "name", connectorName
+                ));
+    }
+
+    @Test
+    public void shouldRetrieveConnectorPlugins() {
+        webTestClient.get()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connector-plugins", clusterName)
+                .exchange()
+                .expectStatus().isOk()
+                .expectBodyList(ConnectorPlugin.class)
+                .value(plugins -> assertEquals(13, plugins.size()));
+    }
+
+    @Test
+    public void shouldSuccessfullyValidateConnectorPluginConfiguration() {
+        var pluginName = "FileStreamSinkConnector";
+        webTestClient.put()
+                .uri("http://localhost:8080/api/clusters/{clusterName}/connector-plugins/{pluginName}/config/validate", clusterName, pluginName)
+                .bodyValue(Map.of(
+                        "connector.class", "org.apache.kafka.connect.file.FileStreamSinkConnector",
+                        "tasks.max", "1",
+                        "topics", "output-topic",
+                        "file", "/tmp/test",
+                        "name", connectorName
+                        )
+                )
+                .exchange()
+                .expectStatus().isOk()
+                .expectBody(ConnectorPluginConfigValidationResponse.class)
+                .value(response -> assertEquals(0, response.getErrorCount()));
+    }
+
+}
